pdf-summarizer/requirements.txt
flask
langchain-community 
langchain-huggingface 
faiss-cpu 
tiktoken 
python-dotenv 
pypdf